pyproject.toml
[tool.poetry]
name = "balancer-rewards-manager"
version = "0.3.0"
description = "Balancer rewards manager"
authors = ["bulbozaur <info.lido.fi>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
eth-brownie = "^1.18.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
